backend/game/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

PyJWT>=2.8.0
requests>=2.31.0
djangorestframework-simplejwt>=5.3.1
django-channels-jwt-auth-middleware>=1.0.0

channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
websockets>=12.0
django-eventstream>=4.5.0
asgiref>=3.7.2

psycopg2-binary>=2.9.9
redis>=5.0.1
django-redis>=5.4.0

django-ratelimit>=4.1.0

python-dotenv>=1.0.0
gunicorn>=21.2.0

django-debug-toolbar>=4.2.0
prometheus-client>=0.19.0

drf-spectacular>=0.27.0

pytest>=7.4.0
pytest-django>=4.7.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

numpy>=1.24.0
pandas>=2.0.0
msgpack>=1.0.5